+++ addOns/graphql/CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Changed
+- The "Import a GraphQL Schema from a File" and "Import a GraphQL schema from a URL" menu items were merged into one,
+  "Import a GraphQL schema".
+- The Import dialog shows the values used in the previous import when reopened.
 
 ## [0.17.0] - 2023-06-19
 ### Added